@@ -2,74 +2,38 @@
   "name": "app",
   "version": "0.1.0",
   "private": true,
-  "engines" : {
-    "node" : ">=16.0.0"
+  "engines": {
+    "node": ">=16.0.0"
   },
   "dependencies": {
-    "@leafygreen-ui/a11y": "^1.3.4",
-    "@leafygreen-ui/badge": "^5.0.1",
-    "@leafygreen-ui/banner": "^4.0.1",
-    "@leafygreen-ui/box": "^3.1.1",
-    "@leafygreen-ui/button": "^15.0.2",
-    "@leafygreen-ui/callout": "^5.0.1",
-    "@leafygreen-ui/card": "^6.0.1",
-    "@leafygreen-ui/checkbox": "^8.0.3",
-    "@leafygreen-ui/code": "^11.0.1",
-    "@leafygreen-ui/combobox": "^1.2.2",
-    "@leafygreen-ui/confirmation-modal": "^3.1.4",
-    "@leafygreen-ui/copyable": "^4.0.2",
-    "@leafygreen-ui/emotion": "^4.0.3",
-    "@leafygreen-ui/expandable-card": "^0.9.6",
-    "@leafygreen-ui/form-footer": "^0.10.3",
-    "@leafygreen-ui/hooks": "^7.4.0",
-    "@leafygreen-ui/icon": "^11.12.4",
-    "@leafygreen-ui/icon-button": "^12.0.0",
-    "@leafygreen-ui/inline-definition": "^4.1.1",
-    "@leafygreen-ui/interaction-ring": "^3.0.0",
-    "@leafygreen-ui/leafygreen-provider": "^2.3.5",
-    "@leafygreen-ui/lib": "^9.5.2",
-    "@leafygreen-ui/logo": "^6.3.0",
-    "@leafygreen-ui/marketing-modal": "^3.0.10",
-    "@leafygreen-ui/menu": "^13.1.1",
-    "@leafygreen-ui/modal": "^9.0.2",
-    "@leafygreen-ui/mongo-menu": "^3.0.3",
-    "@leafygreen-ui/mongo-nav": "^9.4.2",
-    "@leafygreen-ui/palette": "^3.4.7",
-    "@leafygreen-ui/pipeline": "^2.2.1",
-    "@leafygreen-ui/popover": "^8.0.1",
-    "@leafygreen-ui/portal": "^4.0.9",
-    "@leafygreen-ui/radio-box-group": "^8.0.2",
-    "@leafygreen-ui/radio-group": "^8.1.2",
-    "@leafygreen-ui/ripple": "^1.1.8",
-    "@leafygreen-ui/segmented-control": "^2.0.1",
-    "@leafygreen-ui/select": "^5.0.3",
-    "@leafygreen-ui/side-nav": "^9.0.3",
-    "@leafygreen-ui/stepper": "^3.1.6",
-    "@leafygreen-ui/table": "^4.1.1",
-    "@leafygreen-ui/tabs": "^7.0.1",
-    "@leafygreen-ui/testing-lib": "^0.2.1",
-    "@leafygreen-ui/text-area": "^6.2.1",
-    "@leafygreen-ui/text-input": "^9.0.0",
-    "@leafygreen-ui/theme": "^3.0.1",
-    "@leafygreen-ui/toast": "^1.0.1",
-    "@leafygreen-ui/toggle": "^8.0.6",
-    "@leafygreen-ui/tokens": "^1.4.1",
-    "@leafygreen-ui/tooltip": "^7.1.3",
-    "@leafygreen-ui/typography": "^12.0.0",
-    "@types/jest": "^27.5.2",
+    "@iconscout/unicons": "^4.0.8",
+    "@mdi/font": "^7.2.96",
+    "@testing-library/react": "latest",
+    "@types/jest": "latest",
     "@types/node": "^16.18.12",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-router-dom": "^6.8.1",
-    "react-scripts": "4.0.3",
+    "@types/react": "latest",
+    "@types/react-dom": "latest",
+    "animate.css": "^4.1.1",
+    "choices.js": "^10.2.0",
+    "feather-icons": "^4.29.0",
+    "gumshoejs": "^5.1.2",
+    "jarallax": "^2.1.3",
+    "js-datepicker": "^5.18.2",
+    "react": "^18.2.0",
+    "react-dom": "latest",
+    "react-router-dom": "latest",
+    "react-scripts": "latest",
+    "shufflejs": "^6.1.0",
+    "swiper": "4.5.0",
+    "tiny-slider": "^2.9.4",
+    "tobii": "^2.0.0-alpha",
     "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "wow.js": "^1.2.2"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts --openssl-legacy-provider build",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -90,5 +54,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@tailwindcss/forms": "^0.5.3",
+    "resolve-url-loader": "^5.0.0",
+    "sass": "^1.65.1",
+    "tailwindcss": "^3.3.3"
   }
 }
